# ui/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchiveRequest": {
        "properties": {
          "names": {
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Names",
            "type": "array"
          }
        },
        "required": [
          "names"
        ],
        "title": "ArchiveRequest",
        "type": "object"
      },
      "AssetDeleted": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "enum": [
              "workspace",
              "common",
              "examples"
            ],
            "title": "Origin",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "name",
          "workspace",
          "reference",
          "deleted",
          "origin"
        ],
        "title": "AssetDeleted",
        "type": "object"
      },
      "AssetFile": {
        "additionalProperties": false,
        "properties": {
          "absolute_url": {
            "title": "Absolute Url",
            "type": "string"
          },
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "kind": {
            "enum": [
              "image",
              "video",
              "audio",
              "text"
            ],
            "title": "Kind",
            "type": "string"
          },
          "mtime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Mtime"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "description": "Which library it came from: this workspace's own, the `common` one every workspace shares, or a read-only examples tree - why a delete can answer 403.",
            "enum": [
              "workspace",
              "common",
              "examples"
            ],
            "title": "Origin",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "writable": {
            "title": "Writable",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "reference",
          "folder",
          "kind",
          "size",
          "mtime",
          "origin",
          "writable",
          "url"
        ],
        "title": "AssetFile",
        "type": "object"
      },
...
      "AssetList": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetFile"
            },
            "title": "Assets",
            "type": "array"
          },
          "folders": {
            "items": {
              "type": "string"
            },
            "title": "Folders",
            "type": "array"
          },
          "libraries": {
            "items": {
              "$ref": "#/components/schemas/AssetLibraryRoot"
            },
            "title": "Libraries",
            "type": "array"
          },
          "shadowed": {
            "items": {
              "$ref": "#/components/schemas/ShadowedAsset"
            },
            "title": "Shadowed",
            "type": "array"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "workspace",
          "libraries",
          "assets",
          "folders",
          "shadowed"
        ],
        "title": "AssetList",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeepRequest": {
        "properties": {
          "asset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Name to keep it under in the asset library; its own file name when omitted. May name a folder; the kept file's extension is assumed when the name has none",
            "title": "Asset Name"
          },
          "name": {
            "description": "The generated file to keep, as the gallery names it",
            "title": "Name",
            "type": "string"
          },
          "overwrite": {
            "default": false,
            "description": "Replace an asset already under that name",
            "title": "Overwrite",
            "type": "boolean"
          },
          "shared": {
            "default": false,
            "description": "Keep it in the library every workspace under this root shares, rather than in this workspace's own",
            "title": "Shared",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ],
        "title": "KeepRequest",
        "type": "object"
      },
      "Kept": {
        "additionalProperties": false,
        "properties": {
          "linked": {
            "title": "Linked",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "shared": {
            "title": "Shared",
            "type": "boolean"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "name",
          "workspace",
          "linked",
          "shared"
        ],
        "title": "Kept",
        "type": "object"
      },
...
      "Uploaded": {
        "additionalProperties": false,
        "properties": {
          "absolute_url": {
            "title": "Absolute Url",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "shared": {
            "title": "Shared",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "workspace",
          "url",
          "shared"
        ],
        "title": "Uploaded",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Declarative diffusers workflows over HTTP: queue a job, stream its progress, fetch what it saved.",
    "title": "diffusers-workflow",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/assets": {
      "get": {
        "description": "The asset library: the input media an 'asset:' reference names.\n\nReported by reference rather than by path - 'asset:uploads/x.png' is\nwhat a workflow argument carries, and a client that only ever sees\nreferences cannot accidentally write a path that means something\nelse on another machine. Empty, not an error, on a server with no\nlibrary configured: nothing is wrong, there is just nowhere for an\nasset to be.",
        "operationId": "list_assets_api_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assets"
      }
    },
    "/api/assets/archive": {
      "post": {
        "description": "Bundle a multi-file asset selection into one zip - the gallery's\nbulk download, for the input side of it.\n\nResolved down the same search path a run resolves 'asset:' in, so a\nselection spanning the workspace's own library, the shared one and\nan examples tree downloads as one archive; the library-relative name\nis the entry name, which is the name the 'asset:' reference carries.",
        "operationId": "archive_assets_api_assets_archive_post",
        "parameters": [
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Assets"
      }
    },
    "/api/assets/keep": {
      "post": {
        "description": "Keep a generated file as an input asset, under a stable name.\n\nA run's files live under '<workflow>/<run id>/', which is the right\nplace for them and the wrong name to build on: 'latest' moves, and a\npinned run id breaks the moment outputs are pruned. Keeping one\ncopies it into the workspace's asset library, where an 'asset:' name\nstays put - which is what turns a generated still or score into an\ninput later workflows can rely on.\n\nWithin the workspace, so nothing crosses a namespace, and no bytes\ncross the network: a client that had to download and re-upload a\nmulti-gigabyte video to reuse one frame would be paying for the\nround trip twice.",
        "operationId": "keep_output_as_asset_api_assets_keep_post",
        "parameters": [
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Kept"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keep Output As Asset"
      }
    },
    "/api/assets/{name}": {
      "delete": {
        "description": "Permanently remove one file from the asset library.\n\nDeletes from whichever library on the search path holds it, the\nworkspace's own first, so the name deleted is the name 'asset:'\nwould have resolved to. An asset a read-only examples tree brought\nwith it is not this server's to delete - the same 403 a read-only\nprompt or workflow answers with.\n\nNot recoverable, and any workflow still carrying that 'asset:'\nreference stops loading. Without this, everything else that writes\nthe library (uploads, keep) had no counterpart and a mistake could\nonly be cleaned up on the box (T014).",
        "operationId": "delete_asset_api_assets__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDeleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Asset"
      }
    },
...
    "/api/uploads": {
      "post": {
        "description": "Save a browser-picked image, video or audio file into the asset library's\nuploads/ subfolder and hand back the reference a workflow argument\ncan carry.\n\nAn upload is input, so it belongs in the asset library rather than\namong generated output, and the reference handed back is\n'asset:uploads/<name>' - portable, and meaningful in a workflow that\nis saved and rerun later. A server with no asset library configured\nanswers 409. The body is the raw file bytes: no multipart parser\ndependency needed for a single-file upload.\n\n`asset_name` stores it under a name of the caller's choosing -\n'cast/priya-voice.wav' rather than the random one a browser upload\ngets - which is what makes a recurring cast's references readable\nin every workflow that carries them. It may name a folder, is\nconfined to the library the way `keep_output`'s is, and takes the\nuploaded file's extension when it has none of its own. Without it\nthe name stays random, so two uploads of the same file never\ncollide.\n\n`shared` puts it in the library every workspace under this root\nshares rather than in this workspace's own - a recurring cast that\nepisode four, in a workspace of its own, still has to reach.",
        "operationId": "upload_media_api_uploads_post",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Name"
            }
          },
          {
            "in": "query",
            "name": "shared",
            "required": false,
            "schema": {
              "default": false,
              "title": "Shared",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uploaded"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Media"
      }
    },
...
  }
}